--- exam/src/main/res/drawable/testpress_primary_color_stroke.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle" >
-
-    <solid android:color="@null"/>
-
-    <corners android:radius="30dp" />
-
-    <stroke
-        android:width="1dp"
-        android:color="@color/testpress_color_primary" />
-
-</shape>
